feature: add hooks, optimize authPage
