+++ .eslintrc.cjs
@@ -27,7 +27,7 @@ module.exports = {
 		'@typescript-eslint/strict-boolean-expressions': 'off',
 		'@typescript-eslint/no-non-null-assertion': 'off',
 		'@typescript-eslint/no-floating-promises': 'off',
-		'@typescript-eslint/return-await': 'off'
+		'@typescript-eslint/return-await': 'off',
 	},
 	settings: {
 		react: {
