+++ src/pages/auth/index.tsx
@@ -10,13 +10,14 @@ import { setToken } from '@store/user/userSlice';
 import { Button } from '@ui/button';
 import { Input } from '@ui/input/indext';
 
+import { useTimer } from '@hooks/useTimer';
+
 import './style.scss';
 
 export const AuthPage: React.FC = () => {
 	const [phone, setPhone] = useState<string>('');
 	const [code, setCode] = useState<string>('');
-	const [timer, setTimer] = useState<number>(0);
-	const userDispatch = useUserDispatch();
+	const [timer, setTimer] = useTimer();
 
 	const resendCodeClass = classNames({
 		'auth__resend-text--disabled': timer > 0,
@@ -29,6 +30,15 @@ export const AuthPage: React.FC = () => {
 	const onCodeChange = (e: React.ChangeEvent<HTMLInputElement>): void => {
 		setCode(e.target.value);
 	};
+	const sendCode = (): void => {
+		if (timer === 0) {
+			createCode.refetch();
+			setTimer(120);
+		}
+	};
+	const onSignIn = (): void => {
+		userAuth.refetch();
+	};
 
 	const createCode = useQuery({
 		queryKey: ['code'],
@@ -39,26 +49,7 @@ export const AuthPage: React.FC = () => {
 		queryFn: async () => signIn(phone, parseInt(code)),
 	});
 
-	const sendCode = (): void => {
-		if (timer === 0) {
-			createCode.refetch();
-			setTimer(120);
-		}
-	};
-
-	useEffect(() => {
-		const interval = setInterval(() => {
-			if (timer > 0) {
-				setTimer(timer - 1);
-			} else {
-				clearInterval(interval);
-			}
-		}, 1000);
-
-		return () => {
-			clearInterval(interval);
-		};
-	}, [timer]);
+	const userDispatch = useUserDispatch();
 
 	useEffect(() => {
 		if (userAuth.isSuccess) {
@@ -69,13 +60,13 @@ export const AuthPage: React.FC = () => {
 	return (
 		<div className="auth">
 			<h1 className="auth__header-text">Авторизация</h1>
-			<p className="auth__subtitle-text">Введите номер телефона для входа в личный кабинет </p>
+			<p className="auth__subtitle-text">Введите номер телефона для входа в личный кабинет</p>
 			<Input placeholder="Телефон" onChange={onPhoneChange} />
 			{createCode.isSuccess ? (
 				<>
 					<Input placeholder="Проверочный код" onChange={onCodeChange} />
 					{userAuth.isError ? <p className="auth__input-error">Неверный код</p> : ''}
-					<Button onClick={() => userAuth.refetch}>Войти</Button>
+					<Button onClick={onSignIn}>Войти</Button>
 					<p onClick={sendCode} className={resendCodeClass}>
 						Запросить код {timer > 0 ? `повторно можно через ${timer} секунд` : ''}{' '}
 					</p>
