+++ src/ui/input/indext.tsx
@@ -1,9 +1,7 @@
+import type { ComponentProps } from 'react';
 import './style.scss';
 
-interface InputProps {
-	placeholder?: string;
-	onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
-}
+interface InputProps extends ComponentProps<'input'> {}
 
 export const Input: React.FC<InputProps> = ({ placeholder, onChange }) => {
 	return <input type="text" className="input" onChange={onChange} placeholder={placeholder} />;
